requirements.txt
# Add your Python dependencies here
ymoikcom
# Environment variable management

python-dotenv>=1.0.0

# Desktop automation
pyautogui>=0.9.54
pillow>=10.0.0
psutil>=5.9.0
pywin32>=306
pynput

# For example:
# requests>=2.28.0
# pandas>=1.4.0